@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.5",
+    "@angular/animations": "14.0.2",
     "@angular/cdk": "^13.3.5",
-    "@angular/common": "13.3.5",
-    "@angular/compiler": "13.3.5",
-    "@angular/core": "13.3.5",
-    "@angular/forms": "13.3.5",
-    "@angular/platform-browser": "13.3.5",
-    "@angular/platform-browser-dynamic": "13.3.5",
-    "@angular/router": "13.3.5",
+    "@angular/common": "14.0.2",
+    "@angular/compiler": "14.0.2",
+    "@angular/core": "14.0.2",
+    "@angular/forms": "14.0.2",
+    "@angular/platform-browser": "14.0.2",
+    "@angular/platform-browser-dynamic": "14.0.2",
+    "@angular/router": "14.0.2",
     "primeflex": "^3.1.3",
     "primeicons": "^5.0.0",
     "primeng": "^13.3.2",
@@ -30,8 +30,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.4",
     "@angular/cli": "~13.3.4",
-    "@angular/compiler-cli": "13.3.5",
-    "@angular/language-service": "13.3.5",
+    "@angular/compiler-cli": "14.0.2",
+    "@angular/language-service": "14.0.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^12.12.38",
@@ -46,6 +46,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.8.2",
     "tslint": "~6.1.0",
-    "typescript": "^4.4.4"
+    "typescript": "^4.7.4"
   }
 }
