@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SelectTestingComponent } from './select-testing.component';
-import { FormBuilder } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 
 describe('SelectTestingComponent', () => {
   let component: SelectTestingComponent;
@@ -10,7 +10,7 @@ describe('SelectTestingComponent', () => {
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SelectTestingComponent ],
-      providers: [FormBuilder]
+      providers: [UntypedFormBuilder]
     })
     .compileComponents();
   }));
